+++ sickrage/providers/torrent/torrentday.py
@@ -45,9 +45,9 @@ class TorrentDayProvider(TorrentProvider):
         self.cache = TVCache(self, min_time=10)
 
         self.urls.update({
-            'login': '{base_url}/t'.format(base_url=self.urls['base_url']),
-            'search': '{base_url}/V3/API/API.php'.format(base_url=self.urls['base_url']),
-            'download': '{base_url}/download.php/%s/%s'.format(base_url=self.urls['base_url'])
+            'login': '{base_url}/t'.format(**self.urls),
+            'search': '{base_url}/V3/API/API.php'.format(**self.urls),
+            'download': '{base_url}/download.php/%s/%s'.format(**self.urls)
         })
 
         self.categories = {
@@ -85,7 +85,7 @@ class TorrentDayProvider(TorrentProvider):
 
         return True
 
-    def search(self, search_params, search_mode='eponly', epcount=0, age=0, epObj=None):
+    def search(self, search_params, age=0, epObj=None):
         results = []
 
         if not self.login():